ai_service/requirements.txt
ultralytics
opencv-python-headless
pillow
numpy
flask
flask-cors
supabase
python-dotenv
scipy
deepface
tf-keras
requests
